+++ CHANGELOG.md
@@ -4,6 +4,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Key Armoring with custom headers
+```go
+(key *Key) ArmorWithCustomHeaders(comment, version string) (string, error)
+(key *Key) GetArmoredPublicKeyWithCustomHeaders(comment, version string) (string, error)
+```
+- Message armoring with custom headers
+```go
+(msg *PGPMessage) GetArmoredWithCustomHeaders(comment, version string) (string, error)
+```
+
+### Changed
+- Improved key and message armoring testing
+
+### Fixed
+- Public key armoring headers
+
 ## [2.0.1] - 2020-05-01
 ### Security
 - Updated underlying crypto library